facemaxx-mobile/analysis-backend/requirements.txt
opencv-python
mediapipe
insightface
onnxruntime
numpy
fastapi
uvicorn[standard]
pydantic
python-multipart
Pillow
